changed the budget from double to int

+++ BangazonAPI/Models/Department.cs
@@ -18,6 +18,6 @@ namespace BangazonAPI.Models
         public string Name { get; set; }
 
         [Required]
-        public double Budget { get; set; }
+        public int Budget { get; set; }
     }
 }
